Add auth to edit and delete button lesson-12-a

+++ laravel/app/Http/Controllers/QuestionsController.php
@@ -64,7 +64,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question',$question)){
+            abort(403,'Access denied');
+        }
         return view('questions.edit')->with('question',$question);
+
     }
 
     /**
@@ -76,6 +80,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question',$question)){
+            abort(403,'Access denied');
+        }
         $question->update($request->only('title','body'));
         return redirect()->route('questions.index')->with('success','Your Question has been Updated');
     }
@@ -88,6 +95,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question',$question)){
+            abort(403,'Access denied');
+        }
         $question->delete();
         return redirect('/questions')->with('success','Your Question has been Deleted');
     }
